pyproject.toml
[project]
name = "agent-development-kit"
version = "0.1.0"
description = "adk"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "~3.12"
dependencies = [
    "google-adk (>=1.15.1,<2.0.0)",
    "firecrawl-py (>=4.4.0,<5.0.0)",
    "langchain-community (>=0.3.31,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
